+++ lib/screens/authencation/signIn/editBox.dart
@@ -1,5 +1,5 @@
 import 'package:chat_fire/auth/authService.dart';
-import 'package:chat_fire/screens/chat_home/chat.dart';
+import 'package:chat_fire/screens/chatHome/chat.dart';
 import 'package:chat_fire/shared/constant.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
