--- lib/screens/chat_home/recent_chats.dart
@@ -1,22 +1,0 @@
-import 'package:chat_fire/screens/chat_home/userList.dart';
-import 'package:flutter/material.dart';
-
-class RecentChats extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: Container(
-        color: Theme.of(context).accentColor,
-        child: Container(
-          decoration: BoxDecoration(
-              color: Colors.white,
-              borderRadius: BorderRadius.only(
-                topLeft: const Radius.circular(30.0),
-                topRight: const Radius.circular(30.0),
-              )),
-          child: UserList(),
-          ),
-        ),
-    );
-  }
-}
