--- lib/screens/chat_home/userList.dart
@@ -1,24 +1,0 @@
-import 'package:chat_fire/models/user.dart';
-import 'package:chat_fire/screens/chat_home/userTile.dart';
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-
-class UserList extends StatefulWidget {
-  @override
-  _UserListState createState() => _UserListState();
-}
-
-class _UserListState extends State<UserList> {
-  @override
-  Widget build(BuildContext context) {
-
-    final users =Provider.of<List<User>>(context) ?? [];
-
-    return ListView.builder(
-      itemCount: users.length,
-      itemBuilder: (context, index) {
-        return UserTile(user: users[index]);
-      },
-    );
-  }
-}
