+++ pubspec.yaml
@@ -25,7 +25,6 @@ dependencies:
   cupertino_icons: ^0.1.2
   google_sign_in: ^4.0.14
   firebase_auth: ^0.15.1
-  firebase_storage: ^3.0.8
   cloud_firestore: ^0.13.0+1
   font_awesome_flutter: ^8.5.0
   provider: ^4.0.3
